# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Blossom Buddy API",
        "description": "API pour la gestion des plantes",
        "contact": {
            "email": "admin@blossombuddy.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local"
        }
    ],
    "paths": {
...
        "/notifications/test-watering-reminder": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Tester l'envoi d'un rappel d'arrosage pour une plante spécifique",
                "operationId": "7a1ee923010a546f68786a3ec05cd0e3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "plant_id",
                                    "city"
                                ],
                                "properties": {
                                    "plant_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Paris"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Notification de test envoyée",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "notification_data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/send-all-reminders": {
            "post": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Déclencher l'envoi de tous les rappels d'arrosage (admin)",
                "operationId": "3fa40043d8f35b3f5f22eff8080ab862",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "days_ahead": {
                                        "description": "Nombre de jours d'avance pour les rappels",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rappels envoyés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        },
                                        "command_output": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/notifications/my-reminders-preview": {
            "get": {
                "tags": [
                    "Notifications"
                ],
                "summary": "Aperçu des prochains rappels pour l'utilisateur connecté",
                "operationId": "2a6e23d4d5c1a084ad3455816639e235",
                "responses": {
                    "200": {
                        "description": "Liste des prochains rappels",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "plant": {
                                                "type": "object"
                                            },
                                            "reminder_date": {
                                                "type": "string"
                                            },
                                            "urgency": {
                                                "type": "string"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}